Reorganize EditLetterDiff files
- Move EditLetterWordDiff to its own file EditLetterWordDiff.kt
- Move ContentIndex and DiffSegment into EditLetterDiff.kt alongside the interface
- Inline diffBrev() at its call sites (DiffBrevHandler, ApplyEditScriptTest)
- Delete DiffSegments.kt
